Minor code quality work
