+++ src/main/java/org/jabref/logic/importer/fileformat/BibtexParser.java
@@ -226,7 +226,6 @@ public class BibtexParser implements Parser {
             skipWhitespace();
         }
 
-        // Instantiate meta data
         try {
             parserResult.setMetaData(metaDataParser.parse(meta, importFormatPreferences.bibEntryPreferences().getKeywordSeparator()));
         } catch (ParseException exception) {
