+++ src/test/java/org/jabref/testutils/interactive/styletester/StyleTester.fxml
@@ -37,7 +37,7 @@
 <?import org.jabref.gui.icon.JabRefIconView?>
 <?import com.jfoenix.controls.JFXColorPicker?>
 <AnchorPane prefHeight="600.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/8.0.121"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.styletester.StyleTesterView">
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.testutils.interactive.styletester.StyleTesterView">
     <TabPane AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
              AnchorPane.topAnchor="0.0">
         <Tab text="Buttons">
